move focus manager up one view level
